ISSUE #5095 - create selector to get new ticket pins
